ai-backend/requirements.txt
fastapi==0.115.1
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1
pymupdf==1.25.4
python-docx==1.1.2
spacy==3.8.4
openai==0.28.0
requests==2.32.3
loguru==0.7.3
pydantic==2.10.6
typer==0.15.2
httpx==0.28.1
numpy==1.26.4
pymongo[srv]==4.6.1
boto3==1.38.17
scikit-learn==1.6.1
scipy==1.14.1
docx2txt==0.8
sentence-transformers==2.6.1
motor==3.4.0
dnspython==2.6.1
orjson==3.10.7